create: Adds an angular front end project to communicate with the backend spring boot app
The angular front end declares models classes such as Subject, Solution and Category in the /model package to interract with their corresponding entity classes in the spring boot backend.
Component classes such as SubjectComponent updates their .html components with records to display to the users.
The services classes such the MultiService communicate with the backend application to fetch data and upadtes the component classes.
The SubjectDate classe in the ./util package models the Subject entity class in the backend with a single Date field declaration to receive the call from the DateController's  getDates(...) method of the backend.