update: adds a 'search' functionality
declares enum 'key' 'value' type for subject categories.
creates 'search-component' for the search button and moves code to the .html file .
creates service that emits a 'Subject' of type 'BehaviorSubject<boolean>' upon every search operation to determine the display of custom 'page-not-found'.
creates custom 'page-not-found' for unmatcheable searches.
subscribes every component to the searchService.found observable.