update: adds search feature
adds typesscript file 'name-and-id.ts' that maps to search results.
adds method in the search.service.ts that communicates with the backend against search requests.
adds file 'search-processing.component.ts' that processes the search request and routes to the 'search-output.component.ts' for display of search result.
implements the jpa query at the backend to handle search requests and delegates to the controller/MultiController.java.
implements observable 'questionDisplay' that clears the search box once the search request has returned the requested data